refacto: blend background into own file
